requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers
Pillow
openai
torch
torchvision 